+++ lib/abstract_dart.dart
@@ -9,7 +9,10 @@
 //
 library abstract_dart;
 
-export 'src/hierarchy/hierarchy.dart';
-export 'src/implementation/fields.dart';
-export 'src/implementation/groups.dart';
-export 'src/implementation/monoids.dart';
+export 'src/hierarchy/base.dart';
+export 'src/hierarchy/scalar.dart';
+export 'src/implementation/base/fields.dart';
+export 'src/implementation/base/groups.dart';
+export 'src/implementation/base/monoids.dart';
+export 'src/implementation/scalar/algebras.dart';
+export 'src/implementation/scalar/vector_spaces.dart';