+++ lib/src/implementation/base/semigroups.dart
@@ -1,0 +1,1 @@
+// Take a look at monoids & groups to find semigroups.